Move ExtractStreamResp type definition into typings.
