+++ components/log-viewer-webui/client/src/typings/query.ts
@@ -43,6 +43,15 @@ const ExtractJobSearchParams = Type.Object({
     logEventIdx: Type.Number(),
 });
 
+interface ExtractStreamResp {
+    _id: string;
+    begin_msg_ix: number;
+    end_msg_ix: number;
+    is_last_chunk: boolean;
+    path: string;
+    stream_id: string;
+}
+
 interface QueryLoadingStateDescription {
     label: string;
     description: string;
@@ -68,6 +77,8 @@ const QUERY_LOADING_STATE_DESCRIPTIONS
         },
     });
 
+
+export type {ExtractStreamResp};
 export {
     EXTRACT_JOB_TYPE,
     ExtractJobSearchParams,
